requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.2
whitenoise==6.11.0
pandas==2.2.3
yfinance==0.2.54
pykrx==1.0.51
setuptools<81
google-genai==0.3.0
openai==1.57.0
pgvector==0.3.6
beautifulsoup4==4.12.3
finance-datareader
exchange-calendars
lxml
